requirements.txt
slack-bolt>=1.18.0
slack-sdk>=3.26.1
google-cloud-bigquery>=3.13.0
google-generativeai>=0.4.0
python-dotenv>=1.0.0
pytz>=2024.1